chore: move two functions to ValueControl file
