+++ app/src/main/java/com/android/sample/ui/filter/FilterPage.kt
@@ -3,14 +3,10 @@ package com.android.sample.ui.filter
 import android.annotation.SuppressLint
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.rememberScrollState
@@ -18,19 +14,11 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Checkbox
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.RangeSlider
-import androidx.compose.material3.SliderDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateMapOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
@@ -38,16 +26,11 @@ import androidx.compose.ui.unit.dp
 import com.android.sample.R
 import com.android.sample.model.filter.Difficulty
 import com.android.sample.model.filter.FilterPageViewModel
-import com.android.sample.model.filter.FloatRange
-import com.android.sample.resources.C.Dimension.PADDING_16
 import com.android.sample.resources.C.Dimension.PADDING_4
 import com.android.sample.resources.C.Dimension.PADDING_8
 import com.android.sample.resources.C.Dimension.SwipePage.BUTTON_ELEVATION
 import com.android.sample.resources.C.Dimension.SwipePage.BUTTON_RADIUS
 import com.android.sample.resources.C.Tag.Filter.UNINITIALIZED_BORN_VALUE
-import com.android.sample.resources.C.Tag.FilterPage.MAX_ITEM_IN_ROW
-import com.android.sample.resources.C.Tag.FilterPage.SLIDER_COLOR_ACTIVE
-import com.android.sample.resources.C.Tag.FilterPage.SLIDER_COLOR_INACTIVE
 import com.android.sample.resources.C.Tag.FilterPage.TIME_RANGE_MAX
 import com.android.sample.resources.C.Tag.FilterPage.TIME_RANGE_MIN
 import com.android.sample.resources.C.TestTag.FilterPage.TEST_TAG_CATEGORY
@@ -56,11 +39,10 @@ import com.android.sample.resources.C.TestTag.FilterPage.TEST_TAG_TIME_RANGE_SLI
 import com.android.sample.resources.C.TestTag.SwipePage.VIEW_RECIPE_BUTTON
 import com.android.sample.ui.navigation.NavigationActions
 import com.android.sample.ui.navigation.Screen
+import com.android.sample.ui.utils.CheckboxGroup
 import com.android.sample.ui.utils.PlateSwipeScaffold
 import com.android.sample.ui.utils.ValueRangeSlider
-import com.android.sample.ui.utils.reformatTime
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 
 /**
  * Composable function to display the filter page.
@@ -185,76 +167,5 @@ fun FilterBox(
                 )
               }
         }
-      }
-}
-
-/**
- * Composable function to display a group of checkboxes.
- *
- * @param title The title of the group.
- * @param items The list of items to display.
- * @param selectedItem The selected item.
- * @param onItemSelect The function to call when an item is selected.
- * @param testTagPrefix The prefix for the test tag.
- * @param emptyValue The value representing an empty selection.
- */
-@SuppressLint("MutableCollectionMutableState", "StateFlowValueCalledInComposition")
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-fun <T> CheckboxGroup(
-    title: String,
-    items: List<T>,
-    selectedItem: StateFlow<T?>,
-    onItemSelect: (T) -> Unit,
-    testTagPrefix: String,
-    emptyValue: T?
-) {
-  // Map items to checkbox states, observed in LaunchedEffect to sync with selectedItem
-  val itemStates = remember {
-    mutableStateMapOf(*items.associateWith { false }.toList().toTypedArray())
-  }
-
-  // Synchronize item states with selectedItem on updates
-  LaunchedEffect(selectedItem) {
-    itemStates.keys.forEach { itemStates[it] = false }
-    if (selectedItem.value != emptyValue) {
-      itemStates[selectedItem.value!!] = true
-    }
-  }
-
-  Column(
-      modifier = Modifier.padding(PADDING_16.dp),
-      horizontalAlignment = Alignment.CenterHorizontally) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleMedium,
-            modifier = Modifier.align(Alignment.Start))
-
-        FlowRow(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.Center,
-            verticalArrangement = Arrangement.Center,
-            maxItemsInEachRow = MAX_ITEM_IN_ROW) {
-              itemStates.forEach { (item, checked) ->
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                  Spacer(modifier = Modifier.height(PADDING_4.dp))
-
-                  Text(
-                      text = item.toString(),
-                      style = MaterialTheme.typography.bodyMedium,
-                      color = MaterialTheme.colorScheme.onPrimary)
-                  Checkbox(
-                      modifier = Modifier.testTag("${testTagPrefix}Checkbox$item"),
-                      checked = checked,
-                      onCheckedChange = { isChecked ->
-                        // Uncheck all other items and select the new item
-                        itemStates.keys.forEach { itemStates[it] = false }
-                        itemStates[item] = isChecked
-                        onItemSelect(item)
-                      })
-                  Spacer(modifier = Modifier.height(PADDING_4.dp))
-                }
-              }
-            }
       }
 }
